+++ src/app/configuration_manager.cpp
