+++ src/app/include/configuration_manager.h
