+++ src/proc/makefile
@@ -1,0 +1,9 @@
+.PHONY: mkquery mkprocessor
+
+all: mkquery mkprocessor
+
+mkquery:
+	@$(MAKE) -C query
+
+mkprocessor:
+	@$(MAKE) -C processor
