+++ src/proc/processor/include/message_processor.h
@@ -15,12 +15,23 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "ivan_log.h"
-#include "ivan_utils.h"
-#include "redis_locking.h"
-#include "configuration_manager.h"
-#include "scene.h"
-#include "query_helper.h"
+#include "include/ivan_log.h"
+#include "include/ivan_utils.h"
+#include "include/redis_locking.h"
+#include "include/configuration_manager.h"
+
+#include "include/transform_interface.h"
+#include "include/transform_factory.h"
+#include "include/user_device_interface.h"
+#include "include/user_device_factory.h"
+#include "include/scene_interface.h"
+#include "include/scene_factory.h"
+
+#include "include/scene_list_interface.h"
+#include "include/scene_list_factory.h"
+
+#include "include/query_helper.h"
+#include "include/processor_interface.h"
 
 #include "rapidjson/document.h"
 
@@ -29,70 +40,14 @@ limitations under the License.
 #include "aossl/uuid/include/uuid_interface.h"
 #include "aossl/zmq/include/zmq_interface.h"
 
-#ifndef SRC_INCLUDE_MESSAGE_PROCESSOR_H_
-#define SRC_INCLUDE_MESSAGE_PROCESSOR_H_
-
-class ProcessResult {
-  int err_code;
-  std::string err_desc;
-  const char * err_cdesc;
-  std::string return_str;
-  bool success;
-
- public:
-  // Default construct returns a successful process result
-  inline ProcessResult() {
-    err_code = 100;
-    err_desc = "";
-    success = true;
-    return_str = "";
-  }
-
-  // Error constructor
-  inline ProcessResult(int ecode, std::string edesc) {
-    err_code = ecode;
-    err_desc = edesc;
-    success = false;
-    return_str = "";
-  }
-
-  // Success constructor
-  inline ProcessResult(std::string data) {
-    return_str = data;
-    err_code = 100;
-    err_desc = "";
-    success = true;
-  }
-
-  // Set the error
-  inline void set_error(int ecode, std::string edesc) {
-    err_code = ecode;
-    err_desc = edesc;
-    success = false;
-    return_str = "";
-  }
-
-  // Set the error
-  inline void set_error(int ecode, const char * edesc) {
-    err_code = ecode;
-    err_cdesc = edesc;
-    err_desc.assign(err_cdesc);
-    success = false;
-    return_str = "";
-  }
-
-  int get_error_code() {return err_code;}
-  std::string get_error_description() {return err_desc;}
-  void set_return_string(std::string key) {return_str = key;}
-  std::string get_return_string() {return return_str;}
-  bool successful() {return success;}
-};
+#ifndef SRC_PROC_INCLUDE_MESSAGE_PROCESSOR_H_
+#define SRC_PROC_INCLUDE_MESSAGE_PROCESSOR_H_
 
 // The class containing core logic for CrazyIvan
 // Accepts an Scene (assuming it is an inbound message)
 // And performs any and all processing on it,
 // Includes Redis Locks, DB Interactions, and any necessary calculations
-class MessageProcessor {
+class MessageProcessor : public ProcessorInterface {
 // -------------------------Internal Variables------------------------------- //
   Neo4jInterface *n = NULL;
   Neo4jComponentFactory *neo_factory = NULL;
@@ -102,7 +57,11 @@ class MessageProcessor {
   uuidInterface *ugen = NULL;
   QueryHelper *qh = NULL;
   std::string proto_resp = "";
-  Scene *resp_scn = NULL;
+  SceneListInterface *resp_scn = NULL;
+  SceneListFactory slfactory;
+  SceneFactory sfactory;
+  TransformFactory tfactory;
+  UserDeviceFactory udfactory;
 
 // --------------------------Internal Methods-------------------------------- //
   std::string create_uuid();
@@ -114,45 +73,45 @@ class MessageProcessor {
     std::string tran_id, std::string scene_id, int msg_format_type);
   void build_string_response(int msg_type, int err_code, std::string err_msg, \
     std::string tran_id, std::string scene_id, std::string dev_id, \
-    Transform &t, int msg_format_type);
+    TransformInterface *t, int msg_format_type);
 
 // -----------------------------Ping & Kill---------------------------------- //
   // Ping the Crazy Ivan Instance
-  ProcessResult* process_ping_message(Scene *obj_msg);
+  ProcessResult* process_ping_message(SceneListInterface *obj_msg);
 
   // Kill the Crazy Ivan Instance
-  ProcessResult* process_kill_message(Scene *obj_msg);
+  ProcessResult* process_kill_message(SceneListInterface *obj_msg);
 
 // ----------------------------CRUD Support---------------------------------- //
   // Create a new scene
-  ProcessResult* process_create_message(Scene *obj_msg);
+  ProcessResult* process_create_message(SceneListInterface *obj_msg);
 
   // Update the details of a scene entry
-  ProcessResult* process_update_message(Scene *obj_msg);
+  ProcessResult* process_update_message(SceneListInterface *obj_msg);
 
   // Query for scene data
-  ProcessResult* process_retrieve_message(Scene *obj_msg);
+  ProcessResult* process_retrieve_message(SceneListInterface *obj_msg);
 
   // Delete a scene
-  ProcessResult* process_delete_message(Scene *obj_msg);
+  ProcessResult* process_delete_message(SceneListInterface *obj_msg);
 
 // -------------------------------------------------------------------------- //
 // -------------------------Device Registration------------------------------ //
   // Register a device to a scene
   // Predict a coordinate transform for the device
   // and pass it back in the response
-  ProcessResult* process_registration_message(Scene *obj_msg);
+  ProcessResult* process_registration_message(SceneListInterface *obj_msg);
 
   // Remove a device from a scene
   // If we are removing the last user device from a scene,
   // then remove the scene and try to move any paths that pass through the scene
-  ProcessResult* process_deregistration_message(Scene *obj_msg);
+  ProcessResult* process_deregistration_message(SceneListInterface *obj_msg);
 
   // Align an objects transformation with a scene
   // If the object is a member of other scenes:
   //  -If a coordinate system transformation doesn't exist, then create one
   //  -If a coordinate system transformation exists, then update it
-  ProcessResult* process_device_alignment_message(Scene *obj_msg);
+  ProcessResult* process_device_alignment_message(SceneListInterface *obj_msg);
 
  public:
   // Constructor
@@ -164,7 +123,7 @@ class MessageProcessor {
     redis_locks = new RedisLocker(rd);
     ugen = u;
     neo_factory = nf;
-    qh = new QueryHelper(n, neo_factory);
+    qh = new QueryHelper(n, neo_factory, con);
   }
 
   // Destructor
@@ -183,7 +142,7 @@ class MessageProcessor {
   // In the case of a create message, return the key of the created object
   // Otherwise, return a blank string in the case of success,
   // and "-1" in the case of failure
-  inline ProcessResult* process_message(Scene *obj_msg) {
+  inline ProcessResult* process_message(SceneListInterface *obj_msg) {
     // Determine what type of message we have, and act accordingly
     int msg_type = obj_msg->get_msg_type();
     processor_logging->debug("Message Type: " + std::to_string(msg_type));
@@ -212,4 +171,4 @@ class MessageProcessor {
   }
 };
 
-#endif  // SRC_INCLUDE_MESSAGE_PROCESSOR_H_
+#endif  // SRC_PROC_INCLUDE_MESSAGE_PROCESSOR_H_
